+++ src/common/src/traits/mod.rs
@@ -1,4 +1,5 @@
 mod cell;
+mod id_iter;
 mod id_map;
 mod index_controller;
 mod shard_client;
@@ -6,6 +7,7 @@ mod shard_controller;
 mod shard_storage;
 
 pub use cell::*;
+pub use id_iter::*;
 pub use id_map::*;
 pub use index_controller::*;
 pub use shard_client::*;