+++ src/index_event/src/lib.rs
@@ -1,0 +1,35 @@
+use candid::Principal;
+use catalyze_shared::CanisterResult;
+use common::{queries, ShardsIndex};
+use ic_cdk::query;
+use serde_bytes::ByteBuf;
+
+mod calls;
+mod index;
+mod storage;
+
+#[query]
+fn icts_name() -> String {
+    queries::icts_name()
+}
+
+#[query]
+fn icts_version() -> String {
+    queries::icts_version()
+}
+
+// Hacky way to expose the candid interface to the outside world
+#[query(name = "__get_candid_interface_tmp_hack")]
+pub fn __export_did_tmp_() -> String {
+    use candid::export_service;
+    use catalyze_shared::event::{Event, EventEntry, EventFilter};
+
+    export_service!();
+    __export_service()
+}
+
+// Method used to save the candid interface to a file
+#[test]
+pub fn candid() {
+    catalyze_shared::candid::save_candid_file("../../candid/index_event.did", __export_did_tmp_());
+}